Adds (unsigned long) casting to PQResultsTableViewController.

+++ PushQuery/View Controllers/PQResultsTableViewController.m
@@ -102,7 +102,7 @@
     UITableViewCell *cell = [AKGenerics cellWithReuseIdentifier:@"cell" class:[UITableViewCell class] style:UITableViewCellStyleValue1 tableView:tableView atIndexPath:indexPath fromStoryboard:YES];
     id <PQQuestion> question = [self.survey.questions objectAtIndex:indexPath.row];
     cell.textLabel.text = question.text;
-    cell.detailTextLabel.text = [NSString stringWithFormat:@"%lu %@", question.responses.count, [AKGenerics pluralizationForCount:question.responses.count singular:@"response" plural:@"responses"]];
+    cell.detailTextLabel.text = [NSString stringWithFormat:@"%lu %@", (unsigned long)question.responses.count, [AKGenerics pluralizationForCount:question.responses.count singular:@"response" plural:@"responses"]];
     return cell;
 }
